Switching languages from java to python.
During my search for libraries I would use for this project, I found it that merging pdfs wasn't difficult. However, converting powerpoints to pdfs is difficult as there are no free libraries for this and the only other option was to have a program installed on the local device for the conversion. Therefore, i've made the decision to not include the powerpoint conversion and have program just merge pdf files.